requirements.txt
absl-py==2.1.0
accelerate==0.29.2
aiohttp==3.9.4
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.6.0
anyio==4.3.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
bitsandbytes==0.43.1
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.0
dataclasses-json==0.5.14
debugpy==1.8.0
decorator==5.1.1
distro==1.9.0
einops==0.7.0
exceptiongroup==1.2.0
executing==2.0.1
faknow==0.0.3
filelock==3.13.4
frozenlist==1.4.1
fsspec==2024.3.1
gdown==4.6.0
gensim==4.3.2
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
grpcio==1.62.1
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
huggingface-hub==0.22.2
idna==3.7
importlib-metadata==6.8.0
ipykernel==6.27.0
ipython==8.17.2
jedi==0.19.1
jieba==0.42.1
Jinja2==3.1.3
joblib==1.2.0
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.5.0
langchain==0.0.179
langchain-community==0.0.32
langchain-core==0.1.42
langchain-openai==0.1.3
langsmith==0.1.45
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.5.8
networkx==3.2.1
nltk==3.8.1
numexpr==2.10.0
numpy==1.26.4
openai==0.27.7
openapi-schema-pydantic==1.2.4
orjson==3.10.0
packaging==23.2
pandas==2.2.2
parso==0.8.3
pillow==10.3.0
pinecone-client==3.1.0
prompt-toolkit==3.0.41
protobuf==4.25.3
psutil==5.9.6
pure-eval==0.2.2
pyarrow==15.0.2
pydantic==1.10.15
pydantic_core==2.18.1
pydeck==0.8.1b0
Pygments==2.17.2
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.1
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
rfc3986==1.5.0
rich==13.7.1
rpds-py==0.18.0
safetensors==0.4.2
scikit-learn==1.4.2
scipy==1.13.0
serpapi==0.1.5
six==1.16.0
smart-open==7.0.4
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
sphinx-markdown-tables==0.0.17
SQLAlchemy==2.0.29
stack-data==0.6.3
streamlit==1.32.2
sympy==1.12
tenacity==8.2.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
threadpoolctl==3.4.0
tiktoken==0.5.2
tokenizers==0.15.2
toml==0.10.2
toolz==0.12.1
torch==2.2.2
tornado==6.3.3
tqdm==4.66.2
traitlets==5.13.0
transformers==4.39.3
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
watchdog==4.0.0
wcwidth==0.2.12
Werkzeug==3.0.2
wrapt==1.16.0
yarl==1.9.4
zipp==3.18.1
